feat: ChatDetailTemplate라는 공통 컴포넌트 만듦.
공통 컴포넌트 만들고 GroupChatDetailPage.tsx와
PersonalChatDetailPage.tsx 리팩토링
